refactor: 이미지들 Image 컴포넌트로 통합
