+++ frontend/src/pages/DictionaryDetail/index.tsx
@@ -4,6 +4,7 @@ import PottedPlant from 'components/@common/Icons/PottedPlant';
 import ThermometerSnow from 'components/@common/Icons/ThermometerSnow';
 import ThermometerSun from 'components/@common/Icons/ThermometerSun';
 import Warning from 'components/@common/Icons/Warning';
+import Image from 'components/@common/Image';
 import DictInfo from 'components/DictInfo';
 import DictInfoSwitch from 'components/DictInfoSwitch';
 import {
@@ -13,7 +14,6 @@ import {
   HeaderBox,
   ManageInfoBox,
   Name,
-  PlantImage,
   PropBox,
   PropsBox,
   Wrapper,
@@ -56,7 +56,7 @@ const DictionaryDetail = () => {
 
   return (
     <Wrapper>
-      <PlantImage src={image} alt={name} />
+      <Image type="wide" src={image} alt={name} size="300px" />
       <HeaderBox>
         <div>
           <Name>{name}</Name>
